+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Next
+
+* Added `timestamp` on `Block` type, returns a reference to the header's timestamp.
+
 ## [0.9.2](https://github.com/streamingfast/substreams-ethereum/releases/tag/v0.9.2)
 
 * Added `parent` on `CallView` to retrieve the parent call of the current `Call` pointed to by `CallView`